backend/requirements.txt
google-generativeai==0.3.2
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
chromadb==0.4.22
python-multipart==0.0.6
langchain==0.1.4
PyPDF2==3.0.1
httpx==0.26.0
